# PersonalDataManager/src/main/resources/KPIDataAPISwagger.yaml
...
openapi: '3.0.0'
info:
  title: My KPI API
  version: 1.2.0
servers:
  - url: 'https://www.snap4city.org/mypersonaldata/api/v1'
paths:
...
  /kpidata/{kpiId}/activities:
    get:
      tags: [KPI Data Activities]
      summary: Retrieve all activities of a KPI Data
      description: Retrieve all activities of a KPI Data
      security:
        - accessToken: [] 
      parameters:
        - $ref: '#/components/parameters/kpiId'
        - $ref: '#/components/parameters/sourceRequestFilter'
        - $ref: '#/components/parameters/accessTypeFilter'
        - $ref: '#/components/parameters/pageNumber'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/sortDirection'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/first'
        - $ref: '#/components/parameters/last'
        - $ref: '#/components/parameters/lang'
      responses:
        200:
          description: A page or a list of KPI Data Activities
        204:
          description: No Data Found
        401:
          description: Right Exceptions
        400:
          description: Wrong Arguments
  /kpidata/{kpiId}/activities/{activityId}:
    get:
      tags: [KPI Data Activities]
      summary: Retrieve the specific KPI Data Activity
      description: Retrieve the specific KPI Data Activity with the activityId passed in the URL
      security:
        - accessToken: [] 
      parameters:
        - $ref: '#/components/parameters/kpiId'
        - $ref: '#/components/parameters/activityId'
        - $ref: '#/components/parameters/lang'
      responses:
        200:
          description: The KPI Data Activity requested
        204:
          description: No Data Found
        401:
          description: Right Exceptions
  /public/kpidata/{kpiId}/activities/{activityId}:
    get:
      tags: [KPI Data Activities]
      summary: Retrieve the specific public KPI Data Activity
      description: Retrieve the specific public KPI Data Activity with the activityId passed in the URL
      parameters:
        - $ref: '#/components/parameters/kpiId'
        - $ref: '#/components/parameters/activityId'
        - $ref: '#/components/parameters/lang'
      responses:
        200:
          description: The KPI Data Activity requested
        204:
          description: No Data Found
        401:
          description: Right Exceptions
  /public/kpidata/{kpiId}/activities:
    get:
      tags: [KPI Data Activities]
      summary: Retrieve all activities of a public KPI Data
      description: Retrieve all activities of a public KPI Data
      parameters:
        - $ref: '#/components/parameters/kpiId'
        - $ref: '#/components/parameters/sourceRequestFilter'
        - $ref: '#/components/parameters/accessTypeFilter'
        - $ref: '#/components/parameters/pageNumber'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/sortDirection'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/first'
        - $ref: '#/components/parameters/last'
        - $ref: '#/components/parameters/lang'
      responses:
        200:
          description: A page or a list of KPI Data Activities
        204:
          description: No Data Found
        401:
          description: Right Exceptions
        400:
          description: Wrong Arguments
...
components:
  securitySchemes:
    accessToken:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
...
    lang:
      in: query
      name: lang
      schema:
        type: string
      required: false
      description: Language of the user (ISO 639 alpha-2 or alpha-3 language code)
    kpiId:
      in: path
      name: kpiId
      schema:
        type: integer
      required: true
      description: The identifier of the KPI Data
...
    activityId:
      in: path
      name: activityId
      schema:
        type: integer
      required: true
      description: The identifier of the KPI Data Activity
    pageNumber:
      in: query
      name: pageNumber
      schema:
        type: integer
      required: false
      description: Which results page should be retrieved. If not indicated, a list is returned.
    pageSize:
      in: query
      name: pageSize
      schema:
        type: integer
      required: false
      description: Elements in a page
    sortDirection:
      in: query
      name: sortDirection
      schema:
        type: string
        enum: ["asc", "desc"]
      required: false
      description: Sort of the elements in a page
    sortBy:
      in: query
      name: sortBy
      schema:
        type: string
      required: false
      description: On the basis of which data you have to sort the elements
...
    from:
      in: query
      name: from
      schema:
        type: string
      required: false
      description: Filtered results based on start time
    to:
      in: query
      name: to
      schema:
        type: string
      required: false
      description: Filtered results based on end time
    first:
      in: query
      name: first
      schema:
        type: integer
      required: false
      description: Return the first elements in the time interval
    last:
      in: query
      name: last
      schema:
        type: integer
      required: false
      description: Return the last elements in the time interval
...
    sourceRequestFilter:
      in: query
      name: sourceRequestFilter
      schema:
        type: string
      required: false
      description: Filters activities based on the sourceRequestFilter value on the sourceRequest property of the saved activities
    accessTypeFilter:
      in: query
      name: accessTypeFilter
      schema:
        type: string
        enum: ["READ", "WRITE", "DELETE"]
      required: false
      description: Filters activities based on the accessTypeFilter value on the accessTyped property of saved activities